Fix issue #22 enforce to use BASH
Fish isn't posix compatible so is nearly imposible to write a script
that can detect it'sel if is launched by fish or bash and excute
different code paths on fuction of the shell.

Instead, enforce to excute it by bash and avoid the nighmare of trying
to write a shell script that works with every shell that any user could
use by default.

A future improvement, would be to change it to work with dash
(/bin/sh on debian and ubuntu), instead of bash, as dash scripts keep working
with bash but the reverse not (ie, dash not implements bash extensions
but is POSIX compatible )